requirements.txt
aistudio-sdk==0.3.8
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astor==0.8.1
attrdict==2.0.1
attrs==26.1.0
babel==2.18.0
bce-python-sdk==0.9.68
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
chardet==7.4.1
charset-normalizer==3.4.7
click==8.3.2
colorlog==6.10.1
contourpy==1.3.3
crc32c==2.8
cssselect==1.4.0
cssutils==2.11.1
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
et_xmlfile==2.0.0
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
fonttools==4.62.1
fsspec==2026.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.1
idna==3.11
ImageIO==2.37.3
imagesize==2.0.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lap==0.5.13
lazy-loader==0.5
lmdb==2.2.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.35.3
more-itertools==11.0.1
mpmath==1.3.0
mysql-connector-python==9.6.0
narwhals==2.19.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.7.3
paddlepaddle==2.6.2
paddlex==3.4.3
pandas==3.0.2
pdf2docx==0.5.12
pillow==12.2.0
polars==1.39.3
polars-runtime-32==1.39.3
premailer==3.10.0
prettytable==3.17.0
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pypdfium2==5.6.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.2
RapidFuzz==3.14.5
rarfile==4.2
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.49
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
tifffile==2026.3.3
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
ultralytics==8.4.35
ultralytics-thop==2.0.18
urllib3==2.6.3
visualdl==2.5.3
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.8
requests==2.33.1flasgger==0.9.7.1
